third_party/blink/web_tests/external/wpt/css/css-grid/parsing/grid-area-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property grid-row-start value 'span calc(-1)'
  assert_true: 'span calc(-1)' is a supported value for grid-row-start. expected true got false
[FAIL] Property grid-row-start value 'span calc(sibling-index() - 2)'
  assert_equals: expected "span 1" but got "span -1"
Harness: the test ran to completion.
